support list index

+++ object/listobject.cpp
@@ -4,7 +4,7 @@ size_t ListObject::get_size() {
     return lselem.size();
 }
 
-value_t ListObject::get_item(size_t i) {
+value_t ListObject::get_item(int i) {
     return lselem[i];
 }
